next version of StringIntConvert. moved to String package.
